+++ components/function-controller/internal/controllers/serverless/function_reconcile.go
@@ -3,11 +3,11 @@ package serverless
 import (
 	"context"
 	"fmt"
-	"github.com/prometheus/common/log"
-	"k8s.io/apimachinery/pkg/util/wait"
+	"time"
 
 	"github.com/go-logr/logr"
 	"github.com/pkg/errors"
+	"golang.org/x/time/rate"
 	appsv1 "k8s.io/api/apps/v1"
 	autoscalingv1 "k8s.io/api/autoscaling/v1"
 	batchv1 "k8s.io/api/batch/v1"
@@ -15,6 +15,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/tools/record"
+	"k8s.io/client-go/util/workqueue"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
@@ -36,17 +37,15 @@ type FunctionReconciler struct {
 	client      resource.Client
 	recorder    record.EventRecorder
 	config      FunctionConfig
-	gitConfig   GitConfig
 	scheme      *runtime.Scheme
 	gitOperator GitOperator
 }
 
-func NewFunction(client resource.Client, log logr.Logger, config FunctionConfig, gitConfig GitConfig, recorder record.EventRecorder) *FunctionReconciler {
+func NewFunction(client resource.Client, log logr.Logger, config FunctionConfig, recorder record.EventRecorder) *FunctionReconciler {
 	return &FunctionReconciler{
 		client:      client,
 		Log:         log.WithName("controllers").WithName("function"),
 		config:      config,
-		gitConfig:   gitConfig,
 		recorder:    recorder,
 		gitOperator: git.NewGit2Go(),
 	}
@@ -62,6 +61,11 @@ func (r *FunctionReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		Owns(&corev1.Service{}).
 		Owns(&autoscalingv1.HorizontalPodAutoscaler{}).
 		WithOptions(controller.Options{
+			RateLimiter: workqueue.NewMaxOfRateLimiter(
+				workqueue.NewItemExponentialFailureRateLimiter(r.config.GitFetchRequeueDuration*time.Second, 300*time.Second),
+				// 10 qps, 100 bucket size.  This is only for retry speed and its only the overall factor (not per item)
+				&workqueue.BucketRateLimiter{Limiter: rate.NewLimiter(rate.Limit(10), 100)},
+			),
 			MaxConcurrentReconciles: 1, // Build job scheduling mechanism requires this parameter to be set to 1. The mechanism is based on getting active and stateless jobs, concurrent reconciles makes it non deterministic . Value 1 removes data races while fetching list of jobs. https://github.com/kyma-project/kyma/issues/10037
 		}).
 		Complete(r)
@@ -163,26 +167,19 @@ func (r *FunctionReconciler) Reconcile(request ctrl.Request) (ctrl.Result, error
 
 	revision, err := r.syncRevision(instance, gitOptions)
 	if err != nil {
+		errMsg := fmt.Sprintf("Sources update failed: %v", err)
 		if git.IsAuthErr(err) {
-			return r.updateStatusWithoutRepository(ctx, ctrl.Result{
-				RequeueAfter: r.gitConfig.GitFetchRequeueDuration,
-			}, instance, serverlessv1alpha1.Condition{
-				Type:               serverlessv1alpha1.ConditionConfigurationReady,
-				Status:             corev1.ConditionFalse,
-				LastTransitionTime: metav1.Now(),
-				Reason:             serverlessv1alpha1.ConditionReasonGitAuthorizationFailed,
-				Message:            fmt.Sprintf("Authorization to git server failed"),
-			})
+			errMsg = "Authorization to git server failed"
 		}
 
 		return r.updateStatusWithoutRepository(ctx, ctrl.Result{
-			RequeueAfter: r.gitConfig.GitFetchRequeueDuration,
+			Requeue: true,
 		}, instance, serverlessv1alpha1.Condition{
 			Type:               serverlessv1alpha1.ConditionConfigurationReady,
 			Status:             corev1.ConditionFalse,
 			LastTransitionTime: metav1.Now(),
 			Reason:             serverlessv1alpha1.ConditionReasonSourceUpdateFailed,
-			Message:            fmt.Sprintf("Sources update failed: %v", err),
+			Message:            errMsg,
 		})
 	}
 
@@ -229,68 +226,6 @@ func (r *FunctionReconciler) onSourceChange(ctx context.Context, instance *serve
 		Reason:             serverlessv1alpha1.ConditionReasonSourceUpdated,
 		Message:            fmt.Sprintf("Sources %s updated", instance.Name),
 	}, repository, commit)
-}
-
-func (r *FunctionReconciler) syncRevision(instance *serverlessv1alpha1.Function, options git.Options) (string, error) {
-	if instance.Spec.Type == serverlessv1alpha1.SourceTypeGit {
-		backoffCfg := r.gitConfig.Backoff
-		backoff := wait.Backoff{Duration: backoffCfg.Duration, Factor: backoffCfg.Factor,
-			Jitter: backoffCfg.Jitter,
-			Steps:  backoffCfg.Steps,
-			Cap:    backoffCfg.Cap,
-		}
-		log.Info("Starting fetching last commit")
-		var commit = ""
-		//check := func() (done bool, err error) {
-		//	id, err := r.gitOperator.LastCommit(options)
-		//	if err != nil {
-		//		log.Warnf("Unable to get last commit from function: %s, cause: %s", instance.Name, err.Error())
-		//		return false, nil
-		//	}
-		//	commit = id
-		//	return true, nil
-		//}
-		err := wait.ExponentialBackoff(backoff, cloningBackoff(r.gitOperator, options, instance.Name, &commit))
-		if err != nil {
-			return "", errors.Wrap(err, "while getting last version of source code")
-		}
-		return commit, nil
-	}
-	return "", nil
-}
-
-func (r *FunctionReconciler) readGITOptions(ctx context.Context, instance *serverlessv1alpha1.Function) (git.Options, error) {
-	if instance.Spec.Type != serverlessv1alpha1.SourceTypeGit {
-		return git.Options{}, nil
-	}
-
-	var gitRepository serverlessv1alpha1.GitRepository
-	if err := r.client.Get(ctx, client.ObjectKey{Namespace: instance.Namespace, Name: instance.Spec.Source}, &gitRepository); err != nil {
-		return git.Options{}, err
-	}
-
-	var auth *git.AuthOptions
-	if gitRepository.Spec.Auth != nil {
-		var secret corev1.Secret
-		if err := r.client.Get(ctx, client.ObjectKey{Namespace: instance.Namespace, Name: gitRepository.Spec.Auth.SecretName}, &secret); err != nil {
-			return git.Options{}, err
-		}
-		auth = &git.AuthOptions{
-			Type:        git.RepositoryAuthType(gitRepository.Spec.Auth.Type),
-			Credentials: r.readSecretData(secret.Data),
-			SecretName:  gitRepository.Spec.Auth.SecretName,
-		}
-	}
-
-	if instance.Spec.Reference == "" {
-		return git.Options{}, fmt.Errorf("reference has to specified")
-	}
-
-	return git.Options{
-		URL:       gitRepository.Spec.URL,
-		Reference: instance.Spec.Reference,
-		Auth:      auth,
-	}, nil
 }
 
 func (r *FunctionReconciler) readDockerConfig(ctx context.Context, instance *serverlessv1alpha1.Function) (DockerConfig, error) {
